# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Go Vue Template",
        "contact": {},
        "license": {
            "name": "Unknown"
        },
        "version": "0.1"
    },
    "paths": {
...
        "/crud/tokenIssuances": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Gets a list for all entities of the TokenIssuance type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sort by id. Value should be ASC or DESC. eg: ?sort.created_at=DESC",
                        "name": "sort.id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by user_id. Value should be ASC or DESC. eg: ?sort.created_at=DESC",
                        "name": "sort.user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by ip_address. Value should be ASC or DESC. eg: ?sort.created_at=DESC",
                        "name": "sort.ip_address",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by created_at. Value should be ASC or DESC. eg: ?sort.created_at=DESC",
                        "name": "sort.created_at",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modelcrud.APITokenIssuance"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Creates a new TokenIssuance",
                "parameters": [
                    {
                        "description": "Creation parameters",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modelcrud.APICreateTokenIssuanceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modelcrud.APITokenIssuance"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "Soft deletes a range of tokenIssuances by their ids",
                "parameters": [
                    {
                        "description": "List of ids to delete",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modelcrud.IDList"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Hard delete tokenIssuance",
                        "name": "hardDelete",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modelcrud.DeletedCount"
                        }
                    }
                }
            }
        },
        "/crud/tokenIssuances/:id": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Gets a single TokenIssuance entity by their id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TokenIssuance id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modelcrud.APIGetTokenIssuancesResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Updates a single TokenIssuance entity based on their id",
                "parameters": [
                    {
                        "description": "Update parameters",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modelcrud.APIUpdateTokenIssuanceRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "TokenIssuance id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modelcrud.APITokenIssuance"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "Soft deletes a single TokenIssuance entity based on their id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TokenIssuance id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modelcrud.APITokenIssuance"
                        }
                    }
                }
            }
        },
...
        "/crud/users/:id": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Gets a single User entity by their id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modelcrud.APIGetUsersResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Updates a single User entity based on their id",
                "parameters": [
                    {
                        "description": "Update parameters",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modelcrud.APIUpdateUserRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "User id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modelcrud.APIUser"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "Soft deletes a single User entity based on their id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Hard delete user",
                        "name": "hardDelete",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modelcrud.APIUser"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "modelcrud.APICreateTokenIssuanceRequest": {
            "type": "object",
            "properties": {
                "ip_address": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "modelcrud.APIGetTokenIssuancesResponse": {
            "type": "object",
            "properties": {
                "next_offset": {
                    "type": "integer"
                },
                "token_issuances": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/modelcrud.APITokenIssuance"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "modelcrud.APIGetUsersResponse": {
            "type": "object",
            "properties": {
                "next_offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/modelcrud.APIUser"
                    }
                }
            }
        },
...
        "modelcrud.APITokenIssuance": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "description": "uuid",
                    "type": "string"
                },
                "ip_address": {
                    "description": "character varying",
                    "type": "string"
                },
                "user_id": {
                    "description": "uuid",
                    "type": "string"
                }
            }
        },
...
        "modelcrud.APIUpdateTokenIssuanceRequest": {
            "type": "object",
            "properties": {
                "ip_address": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "modelcrud.APIUpdateUserRequest": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "gender_self_defined": {
                    "type": "boolean"
                },
                "image": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sub": {
                    "type": "string"
                }
            }
        },
        "modelcrud.APIUser": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "created_at": {
                    "description": "timestamp without time zone",
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "description": "character varying",
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "gender_self_defined": {
                    "type": "boolean"
                },
                "id": {
                    "description": "uuid",
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "name": {
                    "description": "character varying",
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sub": {
                    "type": "string"
                },
                "updated_at": {
                    "description": "timestamp without time zone",
                    "type": "string"
                }
            }
        },
        "modelcrud.DeletedCount": {
            "type": "object",
            "properties": {
                "deleted_count": {
                    "type": "integer"
                }
            }
        },
        "modelcrud.IDList": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}